requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg[binary]
python-dotenv
pydantic
python-docx
pandas
openpyxl
pypdf
xlrd
python-telegram-bot
